pyproject.toml
[project]
name = "databricks-docs"
dynamic = ["version"]
description = "llmops-databricks-course-VanessaSchl - LLMOps Course on Databricks"
requires-python = ">=3.12, <3.13"
dependencies = [
    "backoff==2.2.1",
    "cffi==1.17.1",
    "cloudpickle==3.1.1",
    "databricks-agents==1.8.2",
    "databricks-mcp==0.9.0",
    "databricks-sdk==0.102.0",
    "databricks-vectorsearch==0.66",
    "feedparser==6.0.11",
    "loguru==0.7.3",
    "mlflow==3.10.1",
    "nest-asyncio==1.6.0",
    "openai==2.8.0",
    "pandas",
    "psycopg==3.3.2",
    "psycopg[binary]==3.3.2",
    "psycopg-pool==3.3.0",
    "pyarrow==22.0.0",
    "pydantic==2.11.7",
    "python-dotenv==1.1.1",
    "trafilatura==2.0.0",
]

[project.optional-dependencies]
dev = [
    "databricks-connect>=17.0, <18",
    "pip>=25.0.1, <26",
    "pre-commit>=4.1.0, <5",
    "claudetracing>=0.2.0, <1",
    "types-pyyaml>=6.0.12.20250915",
]

ci = [
    "pre-commit>=4.1.0, <5",
    "pytest>=8.3.4, <9",
]

[build-system]
requires = ["setuptools>=72.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]
include = ["databricks_docs*"]

[tool.setuptools.dynamic]
version = { file = "version.txt" }

[tool.pytest.ini_options]
addopts = "-s --no-header --no-summary"
testpaths = ["tests"]
pythonpath = ["."]

# ============================================================================
# Ruff Configuration
# ============================================================================

[tool.ruff]
line-length = 90

lint.select = [
    "F",    # pyflakes rules
    "E",    # pycodestyle error rules
    "W",    # pycodestyle warning rules
    "B",    # flake8-bugbear rules
    "I",    # isort rules
    "UP",   # pyupgrade rules
    "SIM",  # flake8-simplify rules
    "ERA",  # eradicate rules
    "C",    # pylint convention rules
    "ANN"   # flake8-annotations rules
]

lint.ignore = [
    "D107",   # Missing docstring in `__init__`
    "D203",   # incorrect-blank-line-before-class
    "D213",   # multi-line-summary-second-line
    "C901",   # too complex
    "ERA001", # commented out code
    "ANN204"
]

[tool.ruff.format]
indent-style = "space"

# ============================================================================
# mypy Configuration
# ============================================================================

[tool.mypy]
python_version = "3.12"

# Plugins
plugins = ["pydantic.mypy"]

# Pydantic plugin settings (optional but recommended)
[tool.pydantic-mypy]
init_forbid_extra = true
init_typed = true
warn_required_dynamic_aliases = true

warn_return_any = true
warn_unused_configs = true
disallow_untyped_decorators = false

strict = true
exclude = [
    "^.azdo/",
    "^.venv/",
    "^resources/",
]

[[tool.mypy.overrides]]
module = "databricks_docs.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "feedparser.*"
ignore_missing_imports = true
